refactor: move per-network staking constants to suite-common/staking
