+++ packages/suite/src/actions/wallet/stake/stakeFormCardanoActions.ts
@@ -10,6 +10,10 @@ import { type DeviceRootState, selectSelectedDevice } from '@suite-common/device
 import { type AdaPools } from '@suite-common/earn-staking-api';
 import { type WithServices } from '@suite-common/redux-utils';
 import {
+    CARDANO_EVERSTAKE_DREP,
+    MIN_CARDANO_AMOUNT_FOR_STAKING,
+    MIN_CARDANO_BALANCE_FOR_STAKING,
+    MIN_CARDANO_FOR_WITHDRAWALS,
     calculate,
     composeStakingTransaction,
     getCardanoAccountPoolId,
@@ -20,14 +24,7 @@ import {
     validateCardanoDrep,
 } from '@suite-common/staking';
 import { notificationsActions } from '@suite-common/toast-notifications';
-import { type NetworkSymbol } from '@suite-common/wallet-config';
-import {
-    CARDANO_EVERSTAKE_DREP,
-    EVERSTAKE_POOL_NAMES,
-    MIN_CARDANO_AMOUNT_FOR_STAKING,
-    MIN_CARDANO_BALANCE_FOR_STAKING,
-    MIN_CARDANO_FOR_WITHDRAWALS,
-} from '@suite-common/wallet-constants';
+import { EVERSTAKE_POOL_NAMES, type NetworkSymbol } from '@suite-common/wallet-config';
 import {
     type AccountVotingDelegation,
     type StakeRootState,
