+++ packages/suite/src/components/earn/modals/EarnInANutshell/components/EarnStakingInfo.tsx
@@ -1,10 +1,13 @@
 import React from 'react';
 
 import { Translation } from '@suite/intl';
-import { getDaysToAddToPoolInitial } from '@suite-common/staking';
+import {
+    CARDANO_ACTIVATION_PERIOD_DAYS,
+    CARDANO_EPOCH_DAYS,
+    getDaysToAddToPoolInitial,
+} from '@suite-common/staking';
 import { EarnFlow } from '@suite-common/suite-types/src/staking';
 import { getNetworkDisplaySymbol } from '@suite-common/wallet-config';
-import { CARDANO_ACTIVATION_PERIOD_DAYS, CARDANO_EPOCH_DAYS } from '@suite-common/wallet-constants';
 import { selectEthValidatorsQueue, selectPoolStatsApy } from '@suite-common/wallet-core';
 import { type Account } from '@suite-common/wallet-types';
 import { StepList } from '@trezor/components';
