+++ packages/suite/src/components/earn/staking/tron/hooks/useTronStakeActions.ts
@@ -4,8 +4,11 @@ import { closeModal, openDeferredModal, preserveModal } from '@suite/modal';
 import { useServices } from '@suite-common/dependency-injection';
 import { useTronStakingStats } from '@suite-common/earn-staking-api';
 import { useDispatch } from '@suite-common/redux-utils';
-import { getTronStakingRewards, getTronWithdrawableBalance } from '@suite-common/staking';
-import { TRON_REPRESENTATIVE_TERMS_OF_SERVICE_URLS } from '@suite-common/wallet-constants';
+import {
+    TRON_REPRESENTATIVE_TERMS_OF_SERVICE_URLS,
+    getTronStakingRewards,
+    getTronWithdrawableBalance,
+} from '@suite-common/staking';
 import {
     type TronFlow,
     type TronStakeError,
