+++ suite-common/staking/src/ethereum/__fixtures__/ethereumStaking.ts
@@ -1,7 +1,7 @@
 import { type EthValidatorsQueue } from '@suite-common/earn-staking-api';
-import { DAYS_TO_ADD_TO_POOL_DEFAULT } from '@suite-common/wallet-constants';
 
 import { ETH_NETWORK_ADDRESSES } from '../ethereumNetworkAddresses';
+import { DAYS_TO_ADD_TO_POOL_DEFAULT } from '../ethereumStakingConstants';
 
 export const transformTxFixtures = [
     {
