+++ suite-common/staking/src/shared/stakingConstants.ts
