+++ suite-common/staking/src/solana/solanaStakingLimitUtils.ts
@@ -1,10 +1,10 @@
-import { WALLET_SDK_SOURCE } from '@suite-common/wallet-constants';
 import { type Account } from '@suite-common/wallet-types';
 import { StakeState } from '@trezor/network-solana/constants';
 import solana from '@trezor/network-solana/runtime';
 import { BigNumber } from '@trezor/utils';
 
 import { getSolanaStakingAccountsByStatus } from './solanaStakingUtils';
+import { WALLET_SDK_SOURCE } from '../shared/stakingConstants';
 
 export type SolanaStakingLimitType = 'claim' | 'unstake';
