+++ suite-common/staking/src/tron/tronStakingConstants.ts
@@ -1,3 +1,5 @@
+import { LUGANODES_TRON_SRS, P2P_ORG_TRON_SRS } from '@suite-common/wallet-config';
+import { type TronTxContractType } from '@suite-common/wallet-constants';
 import { BigNumber } from '@trezor/utils';
 
 export const MIN_TRON_AMOUNT_FOR_STAKING = new BigNumber(1);
@@ -8,11 +10,6 @@ export const TRON_STAKING_RESERVE = new BigNumber(0.5);
 export const MIN_TRON_BALANCE_FOR_STAKING =
     MIN_TRON_AMOUNT_FOR_STAKING.plus(MIN_TRON_FOR_WITHDRAWALS);
 
-// Super Representative addresses offered for voting in Suite
-// source: https://earn.trezor.io/staking/v1/trx/stats
-export const LUGANODES_TRON_SRS = ['TGyrSc9ZmTdbYziuk1SKEmdtCdETafewJ9'];
-export const P2P_ORG_TRON_SRS = ['TH7Fe1W8CcLeqN4LGfqX1R9EpsnrJBQJij'];
-
 // TODO: return these from the API
 export const LUGANODES_TERMS_OF_SERVICE_URL = 'https://luganodes.com/terms-of-service';
 export const P2P_ORG_TERMS_OF_SERVICE_URL = 'https://www.p2p.org/terms-of-use';
@@ -21,3 +18,13 @@ export const TRON_REPRESENTATIVE_TERMS_OF_SERVICE_URLS: Record<string, string> =
     [LUGANODES_TRON_SRS[0] as string]: LUGANODES_TERMS_OF_SERVICE_URL,
     [P2P_ORG_TRON_SRS[0] as string]: P2P_ORG_TERMS_OF_SERVICE_URL,
 };
+
+export const TRON_STAKING_CONTRACT_TYPES: TronTxContractType[] = [
+    'FreezeBalanceContract',
+    'FreezeBalanceV2Contract',
+    'UnfreezeBalanceContract',
+    'UnfreezeBalanceV2Contract',
+    'VoteWitnessContract',
+    'WithdrawExpireUnfreezeContract',
+    'WithdrawBalanceContract',
+];
