+++ suite-common/wallet-config/package.json
@@ -14,7 +14,6 @@
     "dependencies": {
         "@suite-common/earn-stablecoin-defs": "workspace:*",
         "@suite-common/networks": "workspace:*",
-        "@suite-common/wallet-constants": "workspace:*",
         "@trezor/crypto-utils": "workspace:*",
         "@trezor/device-utils": "workspace:*",
         "@trezor/network-cardano": "workspace:*",
