+++ suite-common/wallet-config/src/stakingProviders.ts
@@ -1,10 +1,11 @@
+import { EVERSTAKE_VOTER_PUBKEYS } from '@trezor/network-solana/constants';
+
 import {
     EVERSTAKE_POOLS,
     FIVE_BINARIES_POOLS,
     LUGANODES_TRON_SRS,
     P2P_ORG_TRON_SRS,
-} from '@suite-common/wallet-constants';
-import { EVERSTAKE_VOTER_PUBKEYS } from '@trezor/network-solana/constants';
+} from './stakingProviderConstants';
 
 type StakingProviderId = 'everstake' | 'fivebinaries' | 'luganodes' | 'p2p.org';
