+++ suite-common/wallet-config/tsconfig.json
@@ -4,7 +4,6 @@
     "references": [
         { "path": "../earn-stablecoin-defs" },
         { "path": "../networks" },
-        { "path": "../wallet-constants" },
         { "path": "../../packages/crypto-utils" },
         { "path": "../../packages/device-utils" },
         {
