+++ suite-common/wallet-constants/src/index.ts
@@ -1,9 +1,7 @@
 export * from './formDraft';
 export * from './sendForm';
 export * from './earnConstants';
-export * from './ethereumStakingConstants';
-export * from './tronStakingConstants';
+export * from './ethereumConstants';
 export * from './cardanoConstants';
-export * from './stakingConstants';
 export * from './accountConstants';
-export * from './tronConstants';
+export type * from './tronConstants';
