+++ suite-common/wallet-core/src/send/sendFormEthereumThunks.ts
@@ -5,13 +5,13 @@ import {
 } from '@suite-common/device';
 import { createThunk } from '@suite-common/redux-utils';
 import { type EvmGasParamsGwei } from '@suite-common/schemas/src/evm';
+import { STAKE_GAS_LIMIT_RESERVE } from '@suite-common/staking';
 import { notificationsActions } from '@suite-common/toast-notifications';
 import { getNetwork, getNetworkDisplaySymbol } from '@suite-common/wallet-config';
 import {
     ETH_CONTRACT_CALL_BACKUP_GAS_LIMIT,
     ETH_SPEED_UP_TX_MULTIPLIER,
     ETH_TRANSFER_BACKUP_GAS_LIMIT,
-    STAKE_GAS_LIMIT_RESERVE,
 } from '@suite-common/wallet-constants';
 import {
     type AccountWithNetworkType,
