+++ suite-native/module-earn/src/hooks/staking/useUnstakeForm.ts
@@ -2,12 +2,13 @@ import { useMemo } from 'react';
 import { useSelector } from 'react-redux';
 
 import {
+    UNSTAKE_INTERCHANGES,
+    WALLET_SDK_SOURCE_MOBILE,
     buildUnstakeData,
     getEthereumStakingAddressByType,
     getStakingLimitsByNetworkSymbol,
 } from '@suite-common/staking';
 import { getNetwork } from '@suite-common/wallet-config';
-import { UNSTAKE_INTERCHANGES, WALLET_SDK_SOURCE_MOBILE } from '@suite-common/wallet-constants';
 import { type AccountsRootState, selectAccountByKey } from '@suite-common/wallet-core';
 import { type AccountKey } from '@suite-common/wallet-types';
 import { fromEther, isPositiveBalance } from '@suite-common/wallet-utils';
