+++ suite-native/staking/package.json
@@ -19,7 +19,6 @@
         "@suite-common/staking": "workspace:*",
         "@suite-common/suite-sync": "workspace:*",
         "@suite-common/wallet-config": "workspace:*",
-        "@suite-common/wallet-constants": "workspace:*",
         "@suite-common/wallet-core": "workspace:*",
         "@suite-common/wallet-types": "workspace:*",
         "@suite-common/wallet-utils": "workspace:*",
