+++ suite-native/staking/src/stakeFormSolanaNativeThunks.ts
@@ -1,12 +1,12 @@
 import { type DeviceRootState, selectSelectedDevice } from '@suite-common/device';
 import { createThunk } from '@suite-common/redux-utils';
 import {
+    WALLET_SDK_SOURCE_MOBILE,
     composeSolanaStakingTransaction,
     isSupportedSolStakingNetworkSymbol,
     prepareSolanaStakeTxData,
 } from '@suite-common/staking';
 import { type NetworkSymbol, getNetwork } from '@suite-common/wallet-config';
-import { WALLET_SDK_SOURCE_MOBILE } from '@suite-common/wallet-constants';
 import {
     type AccountsRootState,
     type BlockchainRootState,
