+++ suite-native/staking/src/stakeNativeThunks.test.ts
@@ -2,12 +2,11 @@ import { combineReducers, isFulfilled, isRejected } from '@reduxjs/toolkit';
 
 import { messageSystemInitialState } from '@suite-common/message-system';
 import { mockActionType, mockReducer } from '@suite-common/redux-utils/mocks';
-import { buildStakeData } from '@suite-common/staking';
+import { WALLET_SDK_SOURCE_MOBILE, buildStakeData } from '@suite-common/staking';
 import { type TrezorDevice } from '@suite-common/suite-types';
 import { mockGetIsWindowVisible, mockOnModalCancel } from '@suite-common/suite-types/mocks';
 import { createTestStore } from '@suite-common/test-utils';
 import { asNetworkSymbol } from '@suite-common/wallet-config';
-import { WALLET_SDK_SOURCE_MOBILE } from '@suite-common/wallet-constants';
 import { prepareSendFormReducer } from '@suite-common/wallet-core';
 import {
     type Account,
